Criado ObterDadosRelatorioAcompanhamentoFrequenciaCommand
