+++ src/SME.SR.Application/Commands/Frequencia/AcompanhamentoFrequencia/ObterDadosRelatorioAcompanhamentoFrequenciaCommand.cs
@@ -1,0 +1,15 @@
+﻿using MediatR;
+using SME.SR.Infra;
+
+namespace SME.SR.Application
+{
+    public class ObterDadosRelatorioAcompanhamentoFrequenciaCommand : IRequest<RelatorioFrequenciaIndividualDto>
+    {
+        public FiltroAcompanhamentoFrequenciaJustificativaDto FiltroRelatorio { get; set; }
+
+        public ObterDadosRelatorioAcompanhamentoFrequenciaCommand(FiltroAcompanhamentoFrequenciaJustificativaDto filtroRelatorio)
+        {
+            FiltroRelatorio = filtroRelatorio;
+        }
+    }
+}
